+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Release changelog are available here : https://gitlab.inria.fr/croco-ocean/croco
 
 ### Fixed
 
+- BENCH : Do not put report status to True for reference variant to avoid
+  to mark test passed even if not (Issue #342)
+
 
 ### Changed
 
@@ -25,6 +28,7 @@ Release changelog are available here : https://gitlab.inria.fr/croco-ocean/croco
 ### Contributors on this release
 
 - Contributors already on board : 
+  R. Benshila, S. Le Gac
 
 
 - New contributors :